app/requirements.txt
Django==5.2.5
python-dotenv==1.1.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-allauth==65.10.0
PyJWT==2.10.1
google-auth==2.43.0

# Utility
gunicorn==23.0.0
requests==2.32.4
psycopg2-binary==2.9.9
loguru==0.7.3

# Background tasks
huey==2.5.4
redis==6.4.0

# OpenAI
openai==2.8.1

# Azure
azure-common==1.1.28
azure-core==1.35.0
azure-storage-blob==12.26.0
azure-storage-common==2.1.0
azure-storage-file==2.1.0

# Google LLM
google-genai==1.50.1
Pillow==12.0.0